mcp-server/requirements.txt
mcp>=1.0.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
httpx>=0.26.0
google-api-python-client>=2.111.0
google-auth-oauthlib>=1.2.0
